[DEV-6889] Responding to changes
