+++ index.js
@@ -14,10 +14,11 @@ import ComingSoon from './components/messages/ComingSoon';
 import InformationBoxes from './components/InformationBoxes';
 import SectionTitle from './components/SectionTitle';
 import Carousel from './components/Carousel';
-import PageHeader, { useDynamicStickyClass } from './components/PageHeader';
+import PageHeader from './components/PageHeader';
 import DownloadIconButton from './components/DownloadIconButton';
 import FiscalYearPicker from './components/FiscalYearPicker';
 import ShareIcon from './components/ShareIcon';
+import { useDynamicStickyClass } from './helpers/pageHeaderHelper';
 
 import "./styles/dist.scss";
